Pull most test case from GlobalModelJUnitTest  to GlobalModelJunitTestCase to reuse it. GlobalModelJUnitParallelTest and GlobalModelJUnitTest will extend GlobalModelJunitTestCase  and run those testcase 	   in parallel and sequential way.
       modified:   build.xml
       new file:   src/edu/rice/cs/drjava/model/GlobalModelJUnitParallelTest.java
       modified:   src/edu/rice/cs/drjava/model/GlobalModelJUnitTest.java
       new file:   src/edu/rice/cs/drjava/model/GlobalModelJunitTestCase.java
       modified:   src/edu/rice/cs/drjava/model/GlobalModelTestCase.java
       modified:   src/edu/rice/cs/drjava/model/junit/DefaultJUnitModel.java
       modified:   src/edu/rice/cs/drjava/model/junit/JUnitParallelTestManager.java
       modified:   src/edu/rice/cs/drjava/model/junit/MyJUnitCore.java
       modified:   src/edu/rice/cs/drjava/model/repl/newjvm/InterpreterJVM.java
       modified:   src/edu/rice/cs/util/LogTest.java
