+++ drjava/src/edu/rice/cs/drjava/model/junit/MyJUnitCore.java
@@ -8,6 +8,8 @@ import org.junit.runner.Runner;
 import org.junit.runner.notification.RunListener;
 import org.junit.runner.notification.RunNotifier;
 
+import edu.rice.cs.util.Log;
+
 /**
  * This class is used to run test in parallel.
  * It uses ParallelComputer to parallel test method and test class 
@@ -15,7 +17,8 @@ import org.junit.runner.notification.RunNotifier;
  *
  */
 public class MyJUnitCore extends JUnitCore {
-
+	public static Log _log = new Log("MyJUnitCore.txt", false);
+	
 	private final RunNotifier runNotifier = new RunNotifier();
 
 	/**
@@ -25,6 +28,7 @@ public class MyJUnitCore extends JUnitCore {
 	 * @return  result of running these test case
 	 */
 	public Result parallelRunClasses(RunListener listener, Class<?>[] classes) {
+		_log.log("start parallelRunClasses");
 		Runner runner = Request.classes(new ParallelComputer(true, true), classes).getRunner();
 		Result result = new Result();
 		RunListener resultListener = result.createListener();
