+++ drjava/src/edu/rice/cs/util/LogTest.java
@@ -219,7 +219,7 @@ public class LogTest extends MultiThreadedTestCase {
   private static final int DELAY = 100;
   
   //TODO task
-  private static final Log ltl = new Log("logtest.txt", true);
+  private static final Log ltl = new Log("logtest.txt", false);
   
   /** Attempts to test Log's behavior when called concurrently from several sources.  Spawns NUM_THREADS LogTestThreads 
     * (see above)that wait a random number between 0 and DELAY milliseconds and then log a message.  The function tests
@@ -248,7 +248,6 @@ public class LogTest extends MultiThreadedTestCase {
     Date now = new Date();
     String s0 = fin.readLine();
     Date time0 = log3.parse(s0);
-    //TODOtask
     ltl.log("earlier = " + earlier);
     ltl.log("now     = " + now);
     ltl.log("time0   = " + time0);
@@ -260,7 +259,6 @@ public class LogTest extends MultiThreadedTestCase {
     for (int i = 0; i < NUM_THREADS; i++) {
       String s1 = fin.readLine();
       Date time1 = log3.parse(s1);
-      //TODOtask
       ltl.log("time1   = " + time1);
       assertTrue("Date of message not after 'earlier' and before 'now'", withinTolerance(earlier, time1, now));
       assertTrue("Date of message not after 'previous time' and before 'now'", withinTolerance(time0, time1, now));
